cmd: cm33: include asm/global_data.h for DECLARE_GLOBAL_DATA_PTR
Fixes a build error due to missing macro definition for
`DECLARE_GLOBAL_DATA_PTR` in cmd/cm33.c. This macro is defined
in <asm/global_data.h>, which was not previously included.

Without this header, the compiler treats the macro as an undeclared
identifier, resulting in a type error during build.

Signed-off-by: Tien Nguyen <tien.nguyen.uh@renesas.com>